+++ src/main/java/roomescape/resolver/AuthMember.java
@@ -1,4 +1,4 @@
-package roomescape.controller.api.resolver;
+package roomescape.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
